feat!: expose language metadata as const
Exposing the LanguageMetadata as const with satisfies allows the values of
its members like `languageId` or `fileExtensions` to be exposed statically.
This allows for patterns like
```
type MyFileExtensions = typeof MyFancyDSLLanguageMetaData.fileExtensions
```
Where then MyFileExtensions is not typed to a generic `string[]`
but to `['.myDsl']` which makes it easier to avoid mistakes.

Unfortunately this is a breaking change because it requires the fileExtensions
on the LanguageMetaData to be readonly to work.